+++ ios/Classes/SwiftNoScreenshotPlugin.swift
@@ -10,17 +10,29 @@ public class SwiftNoScreenshotPlugin: NSObject, FlutterPlugin, FlutterStreamHand
     private var eventSink: FlutterEventSink? = nil
     private var lastSharedPreferencesState: String = ""
     private var hasSharedPreferencesChanged: Bool = false
+    
+    private static let ENABLESCREENSHOT = false
+    private static let DISABLESCREENSHOT = true
+
+    private static let preventScreenShotKey = "preventScreenShot"
+    private static let methodChannelName = "com.flutterplaza.no_screenshot_methods"
+    private static let eventChannelName = "com.flutterplaza.no_screenshot_streams"
+    private static let screenshotPathPlaceholder = "screenshot_path_placeholder"
 
     init(screenProtectorKit: ScreenProtectorKit) {
         self.screenProtectorKit = screenProtectorKit
+        super.init()
+
+        // Restore the saved state from UserDefaults
+        var fetchVal = UserDefaults.standard.bool(forKey: SwiftNoScreenshotPlugin.preventScreenShotKey)
+        updateScreenshotState(isScreenshotBlocked: fetchVal)
     }
 
     public static func register(with registrar: FlutterPluginRegistrar) {
-        methodChannel = FlutterMethodChannel(name: "com.flutterplaza.no_screenshot_methods", binaryMessenger: registrar.messenger())
-        eventChannel = FlutterEventChannel(name: "com.flutterplaza.no_screenshot_streams", binaryMessenger: registrar.messenger())
+        methodChannel = FlutterMethodChannel(name: methodChannelName, binaryMessenger: registrar.messenger())
+        eventChannel = FlutterEventChannel(name: eventChannelName, binaryMessenger: registrar.messenger())
         
         let window = UIApplication.shared.delegate?.window
-        
         let screenProtectorKit = ScreenProtectorKit(window: window as? UIWindow)
         screenProtectorKit.configurePreventionScreenshot()
         
@@ -31,36 +43,55 @@ public class SwiftNoScreenshotPlugin: NSObject, FlutterPlugin, FlutterStreamHand
     }
 
     public func applicationWillResignActive(_ application: UIApplication) {
-        updateScreenshotState()
+        persistState()
     }
 
     public func applicationDidBecomeActive(_ application: UIApplication) {
-        updateScreenshotState()
+        fetchPersistedState()
+    }
+    
+    public func applicationWillEnterForeground(_ application: UIApplication) {
+        fetchPersistedState()
+    }
+    
+    public func applicationDidEnterBackground(_ application: UIApplication) {
+        persistState()
     }
 
+    public func applicationWillTerminate(_ application: UIApplication) {
+        persistState()
+    }
+    
+    func persistState() {
+        // Persist the state when changed
+        UserDefaults.standard.set(SwiftNoScreenshotPlugin.preventScreenShot, forKey: SwiftNoScreenshotPlugin.preventScreenShotKey)
+        print("Persisted state: \(SwiftNoScreenshotPlugin.preventScreenShot)")
+        updateSharedPreferencesState("")
+    }
+    
+    func fetchPersistedState() {
+        // Restore the saved state from UserDefaults
+        var fetchVal = UserDefaults.standard.bool(forKey: SwiftNoScreenshotPlugin.preventScreenShotKey) ? SwiftNoScreenshotPlugin.DISABLESCREENSHOT :SwiftNoScreenshotPlugin.ENABLESCREENSHOT
+        updateScreenshotState(isScreenshotBlocked: fetchVal)
+        print("Fetched state: \(SwiftNoScreenshotPlugin.preventScreenShot)")
+    }
+    
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
         switch call.method {
         case "screenshotOff":
-            SwiftNoScreenshotPlugin.preventScreenShot = false
             shotOff()
-            updateSharedPreferencesState("")
             result(true)
         case "screenshotOn":
-            SwiftNoScreenshotPlugin.preventScreenShot = true
             shotOn()
-            updateSharedPreferencesState("")
             result(true)
         case "toggleScreenshot":
-            SwiftNoScreenshotPlugin.preventScreenShot.toggle()
-            !SwiftNoScreenshotPlugin.preventScreenShot ? shotOff() : shotOn()
-            updateSharedPreferencesState("")
+            SwiftNoScreenshotPlugin.preventScreenShot ? shotOn(): shotOff()
             result(true)
         case "startScreenshotListening":
             startListening()
             result("Listening started")
         case "stopScreenshotListening":
             stopListening()
-            updateSharedPreferencesState("")
             result("Listening stopped")
         default:
             result(FlutterMethodNotImplemented)
@@ -68,28 +99,34 @@ public class SwiftNoScreenshotPlugin: NSObject, FlutterPlugin, FlutterStreamHand
     }
 
     private func shotOff() {
+        SwiftNoScreenshotPlugin.preventScreenShot = SwiftNoScreenshotPlugin.DISABLESCREENSHOT
         screenProtectorKit?.enabledPreventScreenshot()
+        persistState()
     }
 
     private func shotOn() {
+        SwiftNoScreenshotPlugin.preventScreenShot = SwiftNoScreenshotPlugin.ENABLESCREENSHOT
         screenProtectorKit?.disablePreventScreenshot()
+        persistState()
     }
 
     private func startListening() {
         NotificationCenter.default.addObserver(self, selector: #selector(screenshotDetected), name: UIApplication.userDidTakeScreenshotNotification, object: nil)
+        persistState()
     }
 
     private func stopListening() {
         NotificationCenter.default.removeObserver(self, name: UIApplication.userDidTakeScreenshotNotification, object: nil)
+        persistState()
     }
 
     @objc private func screenshotDetected() {
         print("Screenshot detected")
-        updateSharedPreferencesState("screenshot_path_placeholder")
+        updateSharedPreferencesState(SwiftNoScreenshotPlugin.screenshotPathPlaceholder)
     }
 
-    private func updateScreenshotState() {
-        if SwiftNoScreenshotPlugin.preventScreenShot {
+    private func updateScreenshotState(isScreenshotBlocked: Bool) {
+        if isScreenshotBlocked {
             screenProtectorKit?.enabledPreventScreenshot()
         } else {
             screenProtectorKit?.disablePreventScreenshot()
@@ -98,7 +135,7 @@ public class SwiftNoScreenshotPlugin: NSObject, FlutterPlugin, FlutterStreamHand
 
     private func updateSharedPreferencesState(_ screenshotData: String) {
         let map: [String: Any] = [
-            "is_screenshot_on": !SwiftNoScreenshotPlugin.preventScreenShot,
+            "is_screenshot_on": SwiftNoScreenshotPlugin.preventScreenShot,
             "screenshot_path": screenshotData,
             "was_screenshot_taken": !screenshotData.isEmpty
         ]